AWS_Sagemaker_Portfolio/Sagemaker-Freelance-Projects/snapCycle/src/requirements.txt
tqdm
tensorflow-gpu
matplotlib
keras